Move WeaponItem to itemtypes package
This is to declutter and better structure the CommonItem module
