+++ CommonItem/src/main/java/module-info.java
@@ -2,4 +2,5 @@ module CommonItem {
 	requires Common;
 	requires CommonWeapon;
 	exports com.github.sef24sp4.common.item;
+	exports com.github.sef24sp4.common.item.itemtypes;
 }
